BotC: Add pytest for pick_from_order
Introducing pick_from_order, a function that takes a list of characters
and a file defining an order and returns the order of the night as
defined by the file, limited to the characters given.

Add beginnings of pytest as a workflow with basic test of this function.
Add pyproject.toml to allow importing files in a parent directory
(thank python for that).