BotC: Add maximum jinxes calculator
Implement maximum jinxes calculator with the ability to treat the search
space of jinxes as a graph, and a basic exhaustive search.

Also add a status check on SIGINT (this would be SIGINFO, but the signal
library doesn't support it). To avoid making an unkillable process, if
a second SIGINT is received in 5 seconds the process will end.